F_T30603 move DocumentBundle Repositories to core
- adjut use statements
