+++ demosplan/DemosPlanCoreBundle/Logic/Document/ParagraphService.php
@@ -17,8 +17,8 @@ use demosplan\DemosPlanCoreBundle\Logic\ApiRequest\EntityFetcher;
 use demosplan\DemosPlanCoreBundle\Logic\CoreService;
 use demosplan\DemosPlanCoreBundle\Logic\DateHelper;
 use demosplan\DemosPlanCoreBundle\Logic\EntityHelper;
-use demosplan\DemosPlanDocumentBundle\Repository\ParagraphRepository;
-use demosplan\DemosPlanDocumentBundle\Repository\ParagraphVersionRepository;
+use demosplan\DemosPlanCoreBundle\Repository\ParagraphRepository;
+use demosplan\DemosPlanCoreBundle\Repository\ParagraphVersionRepository;
 use Doctrine\ORM\NonUniqueResultException;
 use Doctrine\ORM\OptimisticLockException;
 use Doctrine\ORM\ORMException;
