+++ demosplan/DemosPlanDocumentBundle/Tools/ServiceImporter.php
@@ -20,8 +20,8 @@ use demosplan\DemosPlanCoreBundle\Exception\TimeoutException;
 use demosplan\DemosPlanCoreBundle\Exception\VirusFoundException;
 use demosplan\DemosPlanCoreBundle\Logic\Document\ParagraphService;
 use demosplan\DemosPlanCoreBundle\Logic\FileService;
+use demosplan\DemosPlanCoreBundle\Repository\ParagraphRepository;
 use demosplan\DemosPlanCoreBundle\Utilities\DemosPlanTools;
-use demosplan\DemosPlanDocumentBundle\Repository\ParagraphRepository;
 use Doctrine\Persistence\ManagerRegistry;
 use Exception;
 use Monolog\Logger;
